Home

This module has a single function listed below which implements the DataStream interface for Excel files. Note that this modules uses Taro which is dependent upon Java.

ExcelDataStreams.ExcelDataStream — Method.

ExcelDataStream( filename::AbstractString; kwargs... )

Constructor for type which gives access to an Excel file as a DataStream

Arguments

filename::AbstractString : the name of the Excel file to stream

Keyword Arguments

columnindices::Vector{Int} – The indices of columns to use as fields. Defaults to empty.

names::Vector{Symbol} – The names to use for the fields. Defaults to empty.

eltypes::Vector{DataTypes} – The types expected for each field. Conversion will be attemped. Defaults to empty.

sheet::String – Which sheet of the Excel file contains the data of interest. Defaults to sheet1.

skipstart::Int – How many initial rows of the Excel file to skip. Defaults to 0.

nastrings::Vector{String} – Strings to consider to be missing data. Defaults to empty.

source